
\name{robbie-package}
\alias{robbie-package}
\alias{robbie}
\docType{package}
\title{General support functions}
\description{Set of functions for general scripting support.}
\details{
  \tabular{ll}{
    Package: \tab robbie\cr
    Type: \tab package\cr
    Version: \tab 0.1\cr
    Date: \tab 2011-12-21\cr
    License: \tab GPL-3\cr
    LazyLoad: \tab yes\cr
}}
\author{Robbie Morrison <robbie@actrix.co.nz>.}
%  \references{}
\keyword{package}
%  \seealso{}
%  \examples{}

% emacs text editor settings
% local variables:
%   mode: Rd
%   make-backup-files: nil
% end:

