
\name{robbie.getScriptName}
\alias{robbie.getScriptName}
\title{Obtain script name}
\description{Obtain the leaf name of current script.}
\usage{
  robbie.getScriptName()
}
\value{
  The leaf name of the current script.
}
\author{
  Robbie Morrison <robbie@actrix.co.nz>.
}
\examples{
 # obtain leaf name of current script
 script <- robbie.getScriptName()
 script
}

% emacs text editor settings
% local variables:
%   mode: Rd
%   make-backup-files: nil
% end:

